Sponsors (#28)
<!--
Provide a brief, descriptive summary of your changes in the title above.

Use appropriate labels to identify the changes in this request.

Delete this commented-out section before submitting your request.
-->

**Description**\
Provide a clear and concise description of the request here.

**Context**\
Provide a clear and concise explanation for the request here, including
references to any relevant issues (i.e. Fixes #0).

**Verification**\
Provide a clear and concise explanation here for any items not checked
below.

- [ ] I have read this project's code of conduct
- [ ] I have read this project's contributing guidelines
- [ ] I have followed this project's coding standards
- [ ] I have followed this project's documentation standards
- [ ] I have added passing tests to verify changes